feat(cli): Beautiful terminal formatting and refactor
