+++ README.md
@@ -5,35 +5,33 @@ It allows you to produce a historical data series for any given duration (consid
 
 ## Setup and usage
 
-Just install from PyPi as shown below. 
+Just install from PyPi as shown below.
 
 ```bash
 pip install b3-cdi-curve
 ```
 
-Now just import it to your project. The package exposes two methods, **update_db()** and **create_time_series(duration: int)**. 
+Now just import it to your project. The package exposes two methods, **sync_db()** and **create_time_series(duration: int)**.
 Below you can find a basic usage example and a more detailed explanation.
 
 ```python
 import b3cdi
 
 # Create and/or update local DB
-b3cdi.update_db()
+b3cdi.sync_db()
 
 # Create time series for the desired duration
 b3cdi.create_time_series(360)
 ```
+![Creating database](./docs/creating_db.png)
 
-### **update_db**
+### **sync_db**
 
 This method loops through all the dates between the beginning of the series and the last working day looking for the prices on the B3 web site. It then generates and saves a local SQLite database (located on ./output/cdi.db).
-
-It has a 2 seconds minimum delay between each hit to prevent generating too many requests. This means that the first time you run this script, it'll take quite some time to get all the files. 
 
 **The first run takes around two and half hours to build the database (respecting the minimum delay for requests) and it takes around 1.35 GB of disk space.**
 
 After it runs once, it'll check for the last inserted date, so it'll look only for working days after the last update. This keeps the database updated for each time you run the code.
-
 
 ### **create_time_series**
