+++ b3cdi/__init__.py
@@ -1,2 +1,1 @@
-from .b3cdi import update_db
-from .b3cdi import create_time_series
+from .b3cdi import sync_db, create_time_series
